# Conflicts:
#	BChef/.classpath
#	BChef/.gitignore
#	BChef/.settings/org.eclipse.m2e.core.prefs
#	BChef/pom.xml
#	BChef/src/eus/healthit/bchef/core/Principal.java
#	BChef/src/eus/healthit/bchef/core/enums/StatusCode.java
#	BChef/src/eus/healthit/bchef/core/models/Recipe.java
#	BChef/src/eus/healthit/bchef/core/models/RecipeStep.java
#	BChef/src/eus/healthit/bchef/core/models/User.java
#	BChef/src/eus/healthit/bchef/core/repository/implementations/RecipeRepository.java

+++ BChef/src/eus/healthit/bchef/core/models/Ingredient.java
@@ -2,13 +2,13 @@ package eus.healthit.bchef.core.models;
 
 public class Ingredient {
 
-	Long id;
+	Integer id;
 	String name;
 
 	String type;
 	String quantity;
 
-	public Ingredient(long id, String name, String type, String amount) {
+	public Ingredient(int id, String name, String type, String amount) {
 		this.id = id;
 		this.name = name;
 		this.type = type;
